Include Username/Password from .netrc file if available
After the conf files are read, this uses the host name to look up a
Username and Password in the user's ~/.netrc file.  The results are
mapped to the keys glynn expects and merged into the options.  This
allows users to specify passwords for their sites in .netrc instead of
in the code base where they might accidentally be publicised.

Requires the netrc gem.  Fixes #48.