Warn user when the credentials from netrc will be overridden from config
If credentials are found in the netrc file but the user also has a user
name or password set in the site's config file, they get a warning
message informing them that those from the config file take precedence.

I use the HighLine CLI library to show the warning text in yellow.