Offer to save the user-entered login/password into the netrc file
If we're at the password prompt it means this password isn't saved yet,
so for convenience, we show a message offering to save the
login/password to the netrc file.  Because this uses netrc several times
in different parts of the file, I decided to read it at the start of the
file so that that can be reused instead of having to re-read it each
time.

Netrc.default_path is used to display the correct netrc file name
depending on the platform.  HighLine is used to display the prompt and
figure out the user input.